add flow control, error handling, color choose

+++ race/src/pages/GaragePage.tsx
@@ -1,22 +1,23 @@
 import React, { useState, useEffect } from 'react';
-import { Car } from '../types'; // Import Car type definition
-import "../styles/GaragePage.css"
+import { Car } from '../types';
+import "../styles/GaragePage.css";
 
 const GaragePage: React.FC = () => {
     const [cars, setCars] = useState<Car[]>([]);
     const [newCarName, setNewCarName] = useState<string>();
     const [selectedCar, setSelectedCar] = useState<Car>();
+    const [newColor, setNewColor] = useState<string>("#FFFFFF");
 
     useEffect(() => {
         fetchCars();
     }, []);
 
-    // Function to fetch cars from the backend
+    // Function to fetch cars from the backend (runs on port 3000)
     const fetchCars = async () => {
         try {
             const response = await fetch('http://localhost:3000/garage');
             if (!response.ok) {
-                throw new Error('Failed to fetch cars');
+                throw new Error('Bad response');
             }
             const data = await response.json();
             setCars(data);
@@ -28,19 +29,24 @@ const GaragePage: React.FC = () => {
     // Function to add a new car to the garage
     const addCar = async () => {
         try {
-            const response = await fetch('http://localhost:3000/garage', {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({ name: newCarName, color: "#FFFFFF" }),
-            });
+            if(newCarName){
+                const response = await fetch('http://localhost:3000/garage', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({ name: newCarName, color: newColor }),
+                });
 
-            if (!response.ok) {
-                throw new Error('Failed to add car');
+                if (!response.ok) {
+                    throw new Error('Bad response');
+                }
+                // Fetch updated cars after adding the new car
+                fetchCars();
+                console.log(`Added a car {name: ${newCarName}, color: ${newColor}}`)
+            }else{
+                throw new Error('Empty name');
             }
-            // Fetch updated cars after adding the new car
-            fetchCars();
         } catch (error) {
             console.error('Error adding car:', error);
         }
@@ -49,15 +55,18 @@ const GaragePage: React.FC = () => {
     // Function to remove a car from the garage
     const removeCar = async (id: number) => {
         try {
-            const response = await fetch(`http://localhost:3000/garage/${id}`, {
-                method: 'DELETE'
-            });
+            if (id <= 0) {
+                throw new Error(`Must have positive id`);
+            } else {
+                const response = await fetch(`http://localhost:3000/garage/${id}`, {
+                    method: 'DELETE'
+                });
 
-            if (!response.ok) {
-                throw new Error(`Failed to delete car with id ${id}`);
+                if (!response.ok) {
+                    throw new Error(`Bad response with ${id}`);
+                }
+                fetchCars();
             }
-            // Fetch updated cars after adding the new car
-            fetchCars();
         } catch (error) {
             console.error('Error deleting car:', error);
         }
@@ -65,19 +74,23 @@ const GaragePage: React.FC = () => {
 
     const changeCar = async () =>{
         try {
-            const response = await fetch(`http://localhost:3000/garage/${selectedCar?.id}`, {
-                method: 'PUT',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({ name: newCarName, color: "#FFFFFF" }),
-            });
+            if(selectedCar){
+                const response = await fetch(`http://localhost:3000/garage/${selectedCar?.id}`, {
+                    method: 'PUT',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    },
+                    body: JSON.stringify({ name: newCarName, color: newColor}),
+                });
 
-            if (!response.ok) {
-                throw new Error('Failed to change car');
+                if (!response.ok) {
+                    throw new Error('Bad response');
+                }
+                fetchCars();
+            }else{
+                throw new Error('No car selected ');
             }
-            // Fetch updated cars after adding the new car
-            fetchCars();
+
         } catch (error) {
             console.error('Error change car:', error);
         }
@@ -100,6 +113,12 @@ const GaragePage: React.FC = () => {
                         onChange={(e) => setNewCarName(e.target.value)}
                         placeholder="Enter car name"
                     />
+                    <input
+                        type="color"
+                        value={newColor}
+                        onChange={(e) => setNewColor(e.target.value)}
+                    />
+
                     <button onClick={() => changeCar()}>UPDATE</button>
                     <button onClick={addCar}>CREATE</button>
                 </div>
